Fix tests DRY up sylo mocks
