+++ crml/sylo/src/e2ee.rs
@@ -96,57 +96,11 @@ impl<T: Trait> Module<T> {}
 
 #[cfg(test)]
 pub(super) mod tests {
-	use super::*;
-	use codec::{Decode, Encode};
-	use primitives::traits::{Lazy, Verify};
-	use runtime_primitives::traits::{Lazy, Verify};
-	use serde::{Deserialize, Serialize};
-
-	use self::primitives::{Blake2Hasher, H256};
 	use self::sr_io::with_externalities;
-	// The testing primitives are very useful for avoiding having to work with signatures
-	// or public keys. `u64` is used as the `AccountId` and no `Signature`s are requried.
-	use self::runtime_primitives::{
-		testing::{Digest, DigestItem, Header},
-		traits::{BlakeTwo256, IdentityLookup},
-		BuildStorage,
-	};
+	use super::*;
+	use mock::{new_test_ext, Origin, Test};
+	use primitives::H256;
 
-	use vault;
-
-	impl_outer_origin! {
-		pub enum Origin for Test {}
-	}
-
-	#[derive(Encode, Decode, Serialize, Deserialize, Debug)]
-	pub struct Signature;
-
-	impl Verify for Signature {
-		type Signer = H256;
-		fn verify<L: Lazy<[u8]>>(&self, _msg: L, _signer: &Self::Signer) -> bool {
-			true
-		}
-	}
-
-	// For testing the module, we construct most of a mock runtime. This means
-	// first constructing a configuration type (`Test`) which `impl`s each of the
-	// configuration traits of modules we want to use.
-	#[derive(Clone, Eq, PartialEq)]
-	pub struct Test;
-	impl system::Trait for Test {
-		type Origin = Origin;
-		type Index = u64;
-		type BlockNumber = u64;
-		type Hash = H256;
-		type Hashing = BlakeTwo256;
-		type Digest = Digest;
-		type AccountId = H256;
-		type Lookup = IdentityLookup<H256>;
-		type Header = Header;
-		type Event = ();
-		type Log = DigestItem;
-		type Signature = Signature;
-	}
 	impl Trait for Test {
 		type Event = ();
 	}
@@ -156,20 +110,9 @@ pub(super) mod tests {
 	impl inbox::Trait for Test {}
 	impl response::Trait for Test {}
 	impl groups::Trait for Test {}
-	impl vault::Trait for Test {}
 	type E2EE = Module<Test>;
 	type Device = device::Module<Test>;
 	type Response = response::Module<Test>;
-
-	// This function basically just builds a genesis storage key/value store according to
-	// our desired mockup.
-	fn new_test_ext() -> sr_io::TestExternalities<Blake2Hasher> {
-		system::GenesisConfig::<Test>::default()
-			.build_storage()
-			.unwrap()
-			.0
-			.into()
-	}
 
 	#[test]
 	fn should_add_device() {
